requirements.txt
Django==1.8.3
django-debug-toolbar==1.3.2
fake-factory==0.5.2
gunicorn==19.3.0
requests==2.7.0
selenium==2.47.1
sqlparse==0.1.16
django-allauth==0.23.0
django-ckeditor==4.5.1
django-solo==1.1.0
Pillow==2.9.0
django-mathfilters==0.3.0
